+++ BE/src/main/java/roomescape/theme/presentation/dto/response/ThemeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.theme.presentation.dto;
+package roomescape.theme.presentation.dto.response;
 
 import roomescape.theme.domain.Theme;
